Re-model characters and refactor movement
[MERGE animatorTest]
Re-structure characters so that Animator (and NavMesh) are spawned onto
the top level object at runtime, rather than being on the child.  This
also involves removing the Camera gimbal to a separate object.
Refactor CharacterMovement so that it no longer has child classes - any
implementation of movement is sent via Character (or its children).