Create custom Quest Editor
Create a custom editor window which allows for quick generation and
customisation of Quest assets.  New nodes can be added for either
Kill or Travel Objectives, and appropriate values can be assigned, and
these objectives can be linked together in a hierarchy.  All this data
is serialized into the Quest .asset file.
At run-time, the Quest spawns behaviour scripts for all active
objectives which identify when an objective is complete, at which point
any subsequent objectives are started.
The Quest triggers a completion event once all its component objectives
are complete