Add Interaction, including as an Objective
Add in an Interactable class, which can be used as a base class for
objects which the player can interact with.
Created a new InteractObjective, with its own editor node and behaviour
class, which completes when the player interacts with the desired
object.
Created an AbilityMonolith, which unlocks a special ability for the
player on interaction