Merge menu_ui branch
Implement multiple scene setup, so that there is a persistent _Main
scene which contains core objects which should not be destroyed.  This
scene is responsible for transitioning between other scenes, including
fading in and out.
Added a main menu scene, which can be controlled with a gamepad, from
which the player can start a new game or quit.